fix: Circular Error
