+++ apps/workspaces/apis/export_settings/triggers.py
@@ -4,7 +4,7 @@ from fyle_accounting_mappings.models import ExpenseAttribute
 
 from apps.fyle.models import ExpenseGroup
 from apps.tasks.models import TaskLog, Error
-from apps.sage_intacct.tasks import update_last_export_details
+from apps.sage_intacct.actions import update_last_export_details
 from apps.workspaces.models import Configuration, LastExportDetail
 from apps.mappings.models import ImportLog
